Move interfaces to dedicated *.types.ts files
