+++ packages/lists/src/components/Viewport/Viewport.tsx
@@ -3,59 +3,12 @@ import { useState, useRef, useEffect } from 'react';
 import { useDebouncedTimeout } from './useDebouncedTimeout';
 import { usePreviousValue } from './usePreviousValue';
 
-export interface IViewportProps {
-  /**
-   * The height of the scroll container (in pixels).
-   */
-  height: number;
+import { Vector2D, ScrollDirection, Axis, IViewportProps, IViewportState } from './Viewport.types';
 
-  /**
-   * The width of the scroll container (in pixels).
-   */
-  width: number;
-
-  /**
-   * The child component is a function component which receives the scroll container's current scroll state as its only argument.
-   */
-  children: (viewportState: IViewportState) => JSX.Element[] | JSX.Element;
-}
-
-export type Vector2D<T> = [T, T];
-
-export enum Axis {
-  X = 0,
-  Y = 1
-}
-
-export enum ScrollDirection {
-  none,
-  backward,
-  forward
-}
-
-export interface IViewportState {
-  /**
-   * Whether the user is currently scrolling.
-   */
-  isScrolling: boolean;
-
-  /**
-   * Vector representing the distance scrolled from the top left corner, i.e. the point [0, 0], in the form
-   * [x, y] where x denotes the horizontal and y denotes the vertical scroll distance.
-   */
-  scrollDistance: Vector2D<number>;
-
-  /**
-   * Vector representing the current scroll direction in the form [x, y] where x denotes the horizontal and y
-   * denotes the vertical scroll direction.
-   */
-  scrollDirection: Vector2D<ScrollDirection>;
-}
-
-const SCROLL_DISTANCE_ORIGIN: Vector2D<Axis> = [0, 0];
+const SCROLL_DISTANCE_ORIGIN: Vector2D<number> = [0, 0];
 const NO_SCROLL_DIRECTION: Vector2D<ScrollDirection> = [ScrollDirection.none, ScrollDirection.none];
 
-const SCROLLING_RESET_TIMEOUT_IN_MILLISECONDS = 200;
+const STOPPED_SCROLLING_TIMEOUT_IN_MILLISECONDS = 200;
 
 /**
  * Determines the scroll direction based on the given current and previous scroll distance.
@@ -90,7 +43,7 @@ export const Viewport = React.memo(
     });
     const prevViewportState = usePreviousValue(viewportState);
 
-    const [scheduleResetScrollingTimeout, clearScrollingResetTimeout] = useDebouncedTimeout(() => {
+    const [scheduleStoppedScrollingTimeout, clearStoppedScrollingTimeout] = useDebouncedTimeout(() => {
       setViewportState((currentViewportState: IViewportState) => {
         return {
           isScrolling: false,
@@ -98,7 +51,7 @@ export const Viewport = React.memo(
           scrollDirection: NO_SCROLL_DIRECTION
         };
       });
-    }, SCROLLING_RESET_TIMEOUT_IN_MILLISECONDS);
+    }, STOPPED_SCROLLING_TIMEOUT_IN_MILLISECONDS);
 
     useEffect(() => {
       if (scrollContainerRef.current) {
@@ -121,7 +74,7 @@ export const Viewport = React.memo(
             scrollDirection: [scrollDirectionX, scrollDirectionY]
           });
 
-          scheduleResetScrollingTimeout();
+          scheduleStoppedScrollingTimeout();
         };
 
         scrollContainerRef.current.addEventListener<'scroll'>('scroll', onScroll);
@@ -131,7 +84,7 @@ export const Viewport = React.memo(
             scrollContainerRef.current.removeEventListener('scroll', onScroll);
           }
 
-          clearScrollingResetTimeout();
+          clearStoppedScrollingTimeout();
         };
       }
     }, []);
